Time: 0 ms (100.00%) | Memory: 15.2 MB (38.51%) - LeetSync

+++ 1547-destination-city/destination-city.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     string destCity(vector<vector<string>>& paths) {
         string ans;
-        unordered_map<string,int>mp;
+        unordered_map<string,short int>mp;
         for(auto &x: paths){
             mp[x[0]]++;
             mp[x[1]]--;
